requirements.txt
fastapi 
uvicorn 
pydantic
pydantic-core 
starlette 
SQLAlchemy 
alembic
psycopg2-binary 
passlib[bcrypt] 
python-jose[cryptography] 
email-validator 
dnspython  
python-dotenv 
python-multipart 
watchfiles 
httptools








